requirements.txt
sentence-transformers
council-ai
dspy-ai
wget
openai
langchain-community
sentencepiece
ipykernel
psycopg2-binary
